requirements.txt
mysql-connector-python==8.2.0
gradio==4.15.0
pandas==2.1.4
numpy==1.25.2
torch==2.1.2
transformers==4.36.2
sqlparse==0.4.4
python-dotenv==1.0.0
accelerate==0.25.0
bitsandbytes==0.41.3
